move MatchResult into match module
