+++ algobattle/battle_styles/iterated.py
@@ -123,81 +123,6 @@ class Iterated(algobattle.battle_style.BattleStyle):
         solved: int = 0
         attempting: int = 0
 
-        def __int__(self) -> int:
+        @property
+        def score(self) -> float:
             return self.solved
-
-        def __str__(self) -> str:
-            return str(int(self))
-
-        def __repr__(self) -> str:
-            return f"Result(cap={self.cap}, solved={self.solved}, attempting={self.attempting}"
-
-    class MatchResult(algobattle.battle_style.BattleStyle.MatchResult[Result]):
-        def format(self) -> str:
-            table = []
-            table.append(["GEN", "SOL", *range(1, self.rounds + 1), "CAP", "AVG"])
-
-            for matchup, res in self.items():
-                padding = [""] * (self.rounds - len(res))
-                if len(res) == 0:
-                    last_cap = ""
-                    avg = ""
-                else:
-                    last_cap = res[-1].cap
-                    avg = sum(int(r) for r in res) // len(res)
-                table.append([matchup.generator, matchup.solver, *res, *padding, last_cap, avg])
-
-            return "Battle Type: Iterated Battle\n" + format_table(table)
-
-        def calculate_points(self, achievable_points: int) -> dict[Team, float]:
-            """Calculate the number of achieved points.
-
-            Each pair of teams fights for the achievable points among one another.
-            These achievable points are split over all matches, as one run reaching
-            the iteration cap poisons the average number of points.
-
-            Parameters
-            ----------
-            achievable_points : int
-                Number of achievable points.
-
-            Returns
-            -------
-            dict
-                A mapping between team names and their achieved points.
-                The format is {team_name: points [...]} for each
-                team for which there is an entry in match_data and points is a
-                float value. Returns an empty dict if no battle was fought.
-            """
-            points: defaultdict[Team, float] = defaultdict(lambda: 0)
-
-            teams: set[Team] = set()
-            for pair in self.keys():
-                teams = teams.union(set(pair))
-            team_combinations = itertools.combinations(teams, 2)
-
-            if len(teams) == 1:
-                return {teams.pop(): achievable_points}
-
-            rounds = len(next(iter(self.values())))
-            if rounds == 0:
-                return {}
-
-            points_per_round = round(achievable_points / rounds, 1)
-            for pair in team_combinations:
-                for i in range(rounds):
-                    solved1 = self[Matchup(*pair)][i].solved  # pair[1] was solver
-                    solved0 = self[Matchup(*pair[::-1])][i].solved  # pair[0] was solver
-
-                    # Default values for proportions, assuming no team manages to solve anything
-                    points_proportion0 = 0.5
-                    points_proportion1 = 0.5
-
-                    if solved0 + solved1 > 0:
-                        points_proportion0 = solved0 / (solved0 + solved1)
-                        points_proportion1 = solved1 / (solved0 + solved1)
-
-                    points[pair[0]] += round(points_per_round * points_proportion0, 1)
-                    points[pair[1]] += round(points_per_round * points_proportion1, 1)
-
-            return points
